% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFulcrum.R
\name{annotateFulcrum}
\alias{annotateFulcrum}
\title{annotateFulcrum}
\usage{
annotateFulcrum(data, dir = NULL)
}
\arguments{
\item{data}{A single dataframe generated with the joinFulcrum function.}

\item{dir}{The path to the base fulcrum directory, dir/data/raw/fulcrum should contain:
island.csv,
location.csv,
trails.csv}
}
\value{
A single dataframe containing all Fulcrum data sources.
This data frame contains all necessary variables from Fulcrum. It also contains data quality flags.
The variable names match the data dictionary.
}
\description{
\code{annotateFulcrum} Adds additional collection location information to the final Fulcrum dataframe
}
