% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGenotypes.R
\name{checkGenotypes}
\alias{checkGenotypes}
\title{checkGenotypes}
\usage{
checkGenotypes(geno_data, fulc_data, return = FALSE)
}
\arguments{
\item{geno_data}{a genotyping dataframe generated from the \code{procGenotypes} function.}

\item{fulc_data}{a single, joined fulcrum dataframe with all collection data.}

\item{return}{logical, if \code{TRUE} the rows of data for specific flags are returned.}
}
\value{
a list of flagged rows in genptyping and fulcrum dataframes for each flag.
}
\description{
\code{checkGenotypes} checks genotyping data for common errors
}
