% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fix_functions.R
\name{checkJoin}
\alias{checkJoin}
\title{checkJoin}
\usage{
checkJoin(data, return = FALSE)
}
\arguments{
\item{data}{dataframe output of joinFulcrum function}

\item{return}{set to TRUE if flagged rows are to be returned as a dataframe}
}
\value{
a dataframe flagged rows in a dataframe for flag of interest
}
\description{
\code{checkJoin} checks for one flag in joined data
}
