% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fix_functions.R
\name{checkParameters}
\alias{checkParameters}
\title{checkParameters}
\usage{
checkParameters(data, return = FALSE)
}
\arguments{
\item{data}{dataframe output of procFulcrum function}

\item{return}{set to TRUE if flagged rows are to be returned in list format}
}
\value{
a list of flagged rows in dataframe for each of three flags of interest
}
\description{
\code{checkParameters} checks for three flags of raw temperature parameters
}
