% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fix_functions.R
\name{checkProc}
\alias{checkProc}
\title{checkProc}
\usage{
checkProc(data, return = FALSE)
}
\arguments{
\item{data}{dataframe output of procFulcrum function}

\item{return}{set to TRUE if flagged rows are to be returned in list format}
}
\value{
a list of dataframes of flagged rows for each of six flags of interest
}
\description{
\code{checkProc} checks for six flags of duplicated and missing parameters in processed data
}
