% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fulc.R
\name{filter_box}
\alias{filter_box}
\title{filter_box}
\usage{
filter_box(longitude, latitude, coords)
}
\arguments{
\item{longitude}{longitude of collection}

\item{latitude}{latitude of collection}

\item{coords}{coordinates of bounding box}
}
\description{
\code{filter_box} finds lat and long inside bounding box
}
