% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fix_functions.R
\name{fixParameters}
\alias{fixParameters}
\title{fixParameters}
\usage{
fixParameters(
  data,
  substrate_temperature_ids = NULL,
  ambient_temperature_ids = NULL,
  ambient_temperature_run_ids = NULL
)
}
\arguments{
\item{data}{dataframe output of procFulcrum function}

\item{substrate_temperature_ids}{vector of ids of substrate_temperature to edit}

\item{ambient_temperature_ids}{vector of ids of ambient_temperature to edit}

\item{ambient_temperature_run_ids}{vector of ids of ambient_temperature_run to edit}
}
\value{
an edited procFulcrum output (list of dataframes)
}
\description{
\code{fixParameters} fixes three flags of temperature parameters
}
