% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fix_functions.R
\name{fixTemperatures}
\alias{fixTemperatures}
\title{fixTemperatures}
\usage{
fixTemperatures(
  data,
  substrate_temperature_ids = NULL,
  ambient_temperature_ids = NULL,
  ambient_temperature_run_ids = NULL
)
}
\arguments{
\item{data}{list format output of \code{procFulcrum}.}

\item{substrate_temperature_ids}{vector of \emph{fulcrum_id} of rows of \emph{substrate_temperature} to revert to original/raw temperature.}

\item{ambient_temperature_ids}{vector of \emph{fulcrum_id} of rows of \emph{ambient_temperature} to revert to original/raw temperature.}

\item{ambient_temperature_run_ids}{vector of \emph{fulcrum_id} of rows of \emph{ambient_temperature_run} to set \emph{proc_ambient_temperature}, \emph{ambient_humidity} to NA.}
}
\value{
an edited (per passed \emph{fulcrum_id} values) list format output of \code{procFulcrum} to be used in easyfulcrum workflow (passed into \code{joinFulcrum()}).
}
\description{
\code{fixTemperatures} fixes flags (3) regarding raw temperature parameters.
}
\details{
To be run after checkTemperatures, user is advised to select the returned \emph{fulcrum_id} for changing (upon inspection)
from \code{checkTemperatures()} to pass into this function.
flagged substrate temperature: occurs when substrate temperature > 40, \code{procFulcrum}
will automatically convert to celsius (assumes temperature incorrectly in fahrenheit).
flagged ambient temperature: occurs when ambient temperature > 40, \code{procFulcrum}
will automatically convert to celsius (assumes temperature incorrectly in fahrenheit).
flagged ambient run temperature: occurs when ambient humidity and temperature are repeated in subsequent measurements,
this is meant to flag values if probe is stuck, \code{procFulcrum} does not automatically modify anything for this flag.

We advise re-running \code{checkTemperatures()} on the saved output of \code{fixTemperatures()} to ensure that wanted edits have appropriately occurred.
}
