% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fix_functions.R
\name{initial_data_check}
\alias{initial_data_check}
\title{initial_data_check}
\usage{
initial_data_check(data, save = FALSE, return = FALSE)
}
\arguments{
\item{data}{dataframe output of joinFulcrum function}

\item{save}{set to TRUE if flagged rows are to be saved}

\item{return}{set to TRUE if flagged rows are to be returned in list format}
}
\value{
a list of flagged rows in dataframe for each of four flags of interest
}
\description{
\code{initial_data_check} checks for four flags of duplicated and missing parameters in joined data
}
