% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_process_data.R
\name{joinFulcGeno}
\alias{joinFulcGeno}
\title{joinFulcGeno}
\usage{
joinFulcGeno(fulc, geno, blast = NULL)
}
\arguments{
\item{fulc}{a collection data frame output from the \code{procFulcrum} function.}

\item{geno}{a genotyping data frame output from the \code{loadGenotypes} function.}

\item{blast}{OPTIONAL, a blast results data frame output from the \code{procSanger} function.}
}
\value{
A single collection dataframe with variables descriped in the data dictionary.
}
\description{
\code{joinFulcGeno} joins the collection data output from the \code{procFulcrum} function
with the genotyping data output from the \code{loadGenotypes} function.
Blast data output from the \code{procSanger} function can also be joined if desired.
}
