% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFulcrum.R
\name{joinFulcrum}
\alias{joinFulcrum}
\title{joinFulcrum}
\usage{
joinFulcrum(data)
}
\arguments{
\item{data}{A list of processed dataframes generated by the \code{procFulcrum} function.
Valid \code{data} can include only Fulcrum sampling data or all Fulcrum data.}
}
\value{
A single, joined dataframe from the processed dataframes supplied in \code{data} list.
}
\description{
\code{joinFulcrum} joins the processed Fulcrum dataframes from the procFulcrum function,
 selects latitude, longitude, and altitude methods with simple priority algorithm.
}
