% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fulc.R
\name{load_fulc}
\alias{load_fulc}
\title{load_fulc}
\usage{
load_fulc(dir)
}
\arguments{
\item{dir}{The path of the directory with five Fulcrum .csv files:
nematode_field_sampling.csv,
nematode_field_sampling_sample_photo.csv,
nematode_isolation.csv,
nematode_isolation_s_labeled_plates.csv,
nematode_isolation_photos.csv}
}
\value{
a list of five data frames generated from the .csv files.
Each data frame has a shortened name.
\tabular{ll}{
collection \tab nematode_field_sampling.csv\cr
collection_photo \tab nematode_field_sampling_sample_photo.csv\cr
isolation \tab nematode_isolation.csv\cr
isolation_slab \tab nematode_isolation_s_labeled_plates.csv\cr
isolation_photo \tab nematode_isolation_photos.csv\cr
}
}
\description{
\code{load_fulc} loads .csv files exported from Fulcrum into R
}
