% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{loadExampleFiles}
\alias{loadExampleFiles}
\title{loadExampleFiles}
\usage{
loadExampleFiles(startdir, projectdirname)
}
\arguments{
\item{startdir}{the working folder where project files will lie.}

\item{projectdirname}{the name of the project, folder will be constructed under \code{startdir}.}
}
\description{
\code{loadExampleFiles} Copies raw .csv and .jpg files from package into example directory.
}
