% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_process_data.R
\name{loadGenotypes}
\alias{loadGenotypes}
\title{loadGenotypes}
\usage{
loadGenotypes(gsKey)
}
\arguments{
\item{gsKey}{a vector containing the google sheet keys for the genotype data to load.
Your Google Sheet keys are found in your Google sheets URL. Select the string
found between the slashes after spreadsheets/d in your Google Sheet URL.}
}
\value{
A dataframe generated from the Google sheets \code{gsKey} argument. If multiple Google sheets
are provided the data are appended using rbind. A \code{project_id} variable is assigned from the Google sheets name.
The genotyping sheet must be named with the convention {YOUR_PROJECT_NAME}_wild_isolate_genotyping.
}
\description{
\code{loadGenotypes} loads genotyping data from project specific genotyping Google sheets
}
