% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{makeDirStructure}
\alias{makeDirStructure}
\title{makeDirStructure}
\usage{
makeDirStructure(startdir, projectdirname)
}
\arguments{
\item{startdir}{the working folder where project files will lie.}

\item{projectdirname}{the name of the project, folder will be constructed under \code{startdir}.}
}
\description{
\code{makeDirStructure} Makes the nested folder directory for use in easyfulcrum.
}
