% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpSheet.R
\name{makeSpSheet}
\alias{makeSpSheet}
\title{makeSpSheet}
\usage{
makeSpSheet(
  data,
  target_sp = c("Caenorhabditis briggsae", "Caenorhabditis elegans",
    "Caenorhabditis tropicalis"),
  dir = NULL
)
}
\arguments{
\item{data}{a checked dataframe generated from the \code{procPhotos}
function.}

\item{target_sp}{a vector of target species names with full genus and species
names. Default target species names are: Caenorhabditis briggsae,
Caenorhabditis elegans, Caenorhabditis tropicalis.}

\item{dir}{Default is NULL.Supply the path to the base fulcrum directory to
export the species sheet to \code{reports/spSheet.csv}.}
}
\value{
a list of dataframes containing species sheets for target species
  with flags for identified issues.
}
\description{
\code{makeSpSheet} produces a species sheet .csv file for target species
}
