% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fix_functions.R
\name{parameter_check}
\alias{parameter_check}
\title{parameter_check}
\usage{
parameter_check(data, save = FALSE, return = FALSE)
}
\arguments{
\item{data}{dataframe output of procFulcrum function}

\item{save}{set to TRUE if flagged rows are to be saved}

\item{return}{set to TRUE if flagged rows are to be returned in list format}
}
\value{
a list of flagged rows in dataframe for each of three flags of interest
}
\description{
\code{parameter_check} checks for three flags of temperature parameters
}
