% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fulc.R
\name{procFulc}
\alias{procFulc}
\title{procFulc}
\usage{
procFulc(dir)
}
\arguments{
\item{dir}{The path of the directory with five Fulcrum .csv files:
nematode_field_sampling.csv,
nematode_field_sampling_sample_photo.csv,
nematode_isolation.csv,
nematode_isolation_s_labeled_plates.csv,
nematode_isolation_photos.csv}
}
\value{
A dataframe \code{fulc_data} generated from the .csv files.
This data frame contains all variables used from the Fulcrum exports
}
\description{
\code{procFulc} loads Fulcrum .csv files, joins collection and isolation data, and adds collection variables
}
