% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFulcrum.R
\name{procFulcrum}
\alias{procFulcrum}
\title{procFulcrum}
\usage{
procFulcrum(data)
}
\arguments{
\item{data}{A list of dataframes generated by the \code{readFulcrum} function.}
}
\value{
A list of five named data frames.
\tabular{ll}{
nematode_field_sampling \tab a processed dataframe from the nematode_field_sampling.csv\cr
nematode_field_sampling_sample_photo \tab a processed dataframe from the nematode_field_sampling_sample_photo.csv\cr
nematode_isolation \tab a processed dataframe from the nematode_isolation.csv\cr
nematode_isolation_s_labeled_plates \tab a processed dataframe from the nematode_isolation_s_labeled_plates.csv\cr
nematode_isolation_photos \tab a processed dataframe from the nematode_isolation_photos.csv\cr
}
}
\description{
\code{procFulcrum} processes raw Fulcrum data read into R with the \code{readFulcrum} function.
Processing consists of dropping unused variables, renaming variables for joining, and adding flags to collection data.
}
