% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procGenotypes.R
\name{procGenotypes}
\alias{procGenotypes}
\title{procGenotypes}
\usage{
procGenotypes(data)
}
\arguments{
\item{data}{a dataframe resulting from \code{readGenotypes} function.}
}
\value{
A dataframe with flags for common errors in Genotyping dataframe.
}
\description{
\code{procGenotypes} adds flags to genotyping dataframe
}
