% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFulcrum.R
\name{readFulcrum}
\alias{readFulcrum}
\title{readFulcrum}
\usage{
readFulcrum(dir)
}
\arguments{
\item{dir}{The path of the directory with five Fulcrum .csv files:
nematode_field_sampling.csv,
nematode_field_sampling_sample_photo.csv,
nematode_isolation.csv,
nematode_isolation_s_labeled_plates.csv,
nematode_isolation_photos.csv}
}
\value{
A list of five named data frames generated from the .csv files.
\tabular{ll}{
nematode_field_sampling \tab nematode_field_sampling.csv\cr
nematode_field_sampling_sample_photo \tab nematode_field_sampling_sample_photo.csv\cr
nematode_isolation \tab nematode_isolation.csv\cr
nematode_isolation_s_labeled_plates \tab nematode_isolation_s_labeled_plates.csv\cr
nematode_isolation_photos \tab nematode_isolation_photos.csv\cr
}
}
\description{
\code{readFulcrum} reads .csv files exported from Fulcrum into R
}
