% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fix_functions.R
\name{checkGenotypes}
\alias{checkGenotypes}
\title{checkGenotypes}
\usage{
checkGenotypes(
  geno_data,
  fulc_data,
  return_geno = TRUE,
  return_flags = FALSE,
  target_sp = c("Caenorhabditis briggsae", "Caenorhabditis elegans",
    "Caenorhabditis tropicalis")
)
}
\arguments{
\item{geno_data}{a genotyping dataframe generated from the \code{readGenotypes} function.}

\item{fulc_data}{a single, joined fulcrum dataframe with all collection data.}

\item{return_geno}{logical, if \code{TRUE} the genotyping data is returned.}

\item{return_flags}{logical, if \code{TRUE} the rows of data for specific flags are returned.}

\item{target_sp}{vector of target species for species id checks. Default target species names are:
Caenorhabditis briggsae, Caenorhabditis elegans, Caenorhabditis tropicalis.}
}
\value{
a list of flagged rows in genotyping and fulcrum dataframes for each flag.
}
\description{
\code{checkGenotypes} checks genotyping data for common errors in \code{readGenotypes} output.
}
