% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGenotypes.R
\name{checkGenotypes}
\alias{checkGenotypes}
\title{checkGenotypes}
\usage{
checkGenotypes(data, return = FALSE)
}
\arguments{
\item{data}{a genotyping dataframe generated from the \code{readGenotypes} function.}

\item{return}{logical, if \code{TRUE} the rows of data for specific flags are returned.}
}
\value{
a list of flagged rows in genptyping dataframe for each flag.
}
\description{
\code{checkGenotypes} checks genotyping data for common errors
}
