% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{generateReport}
\usage{
generateReport(data, dir)
}
\arguments{
\item{data}{a checked dataframe generated from the \code{procPhotos} function.}

\item{dir}{The path to the base fulcrum directory, report html will be saved as reports/spSheet.csv}
}
\value{
a report as an html file
}
\description{
\code{generateReport} produces a report html for project
}
