% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFulcrum.R
\name{joinFulcrum}
\alias{joinFulcrum}
\title{joinFulcrum}
\usage{
joinFulcrum(data)
}
\arguments{
\item{data}{A list of dataframes generated by the \code{procFulcrum} function.}
}
\value{
A single dataframe containing all Fulcrum data sources.
}
\description{
\code{joinFulcrum} joins the processed Fulcrum dataframes from the procFulcrum function, selects latitude, longitude, and altitude methods with simple priority algorithm.
}
