% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFulcrum.R
\name{joinFulcrum}
\alias{joinFulcrum}
\title{joinFulcrum}
\usage{
joinFulcrum(data, select_vars = T)
}
\arguments{
\item{data}{A list of processed dataframes generated by the
\code{procFulcrum} function. Valid \code{data} can include only Fulcrum
filed sampling data or all Fulcrum data including field sampling and
isolation data.}

\item{select_vars}{Logical, TRUE  will return only the default variables,
FALSE will return all variables. FALSE is recommended if using customized
Fulcrum applications other than "Nematode field sampling" and "Nematode
isolation". TRUE is default.}
}
\value{
A single, joined dataframe from the processed dataframes supplied in
  the \code{data} list.
}
\description{
\code{joinFulcrum} joins the processed Fulcrum dataframes from the
procFulcrum function, and selects latitude, longitude, and altitude methods
with simple priority algorithm.
}
