% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinGenoFulc.R
\name{joinGenoFulc}
\alias{joinGenoFulc}
\title{joinGenoFulc}
\usage{
joinGenoFulc(geno, fulc, blast = NULL)
}
\arguments{
\item{geno}{a genotyping data frame output from the \code{loadGenotypes} function.}

\item{fulc}{a collection data frame output from the \code{procFulcrum} function.}

\item{blast}{OPTIONAL, a blast results data frame output from the \code{procSanger} function.}
}
\value{
A single collection dataframe with variables descriped in the data dictionary.
}
\description{
\code{joinGenoFulc} joins the collection data output from the \code{procFulcrum} function
with the genotyping data output from the \code{readGenotypes} function.
Blast data output from the \code{procSanger} function can also be joined if desired.
}
