% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{makeDirStructure}
\alias{makeDirStructure}
\title{makeDirStructure}
\usage{
makeDirStructure(startdir, projectdirname = NULL)
}
\arguments{
\item{startdir}{the system directory where project directory will be created.}

\item{projectdirname}{the name of the project, a new directory named after
the project will be made under \code{startdir}. If \code{NULL} only the
subdirectories will be made under \code{startdir}.}
}
\description{
\code{makeDirStructure} Makes the required directory structure for use with
\code{easyfulcrum}.
}
