% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_process_data.R
\name{procFulcrum}
\alias{procFulcrum}
\title{procFulcrum}
\usage{
procFulcrum(dir)
}
\arguments{
\item{dir}{The path of the directory with five Fulcrum .csv files:
nematode_field_sampling.csv,
nematode_field_sampling_sample_photo.csv,
nematode_isolation.csv,
nematode_isolation_s_labeled_plates.csv,
nematode_isolation_photos.csv}
}
\value{
A dataframe generated from the .csv files.
This data frame contains all necessary variables from Fulcrum. It also contains data quality flags. The variable names match the data dictionary.
}
\description{
\code{procFulcrum} loads Fulcrum .csv files, joins collection and isolation data, and adds collection variables
}
