% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procPhotos.R
\name{procPhotos}
\alias{procPhotos}
\title{procPhotos}
\usage{
procPhotos(dir, data, percentage = 20, overwrite = FALSE)
}
\arguments{
\item{dir}{a directory with sample photos.}

\item{data}{a data frame output from the \code{joinGenoFulc} function.}

\item{percentage}{This value sets the size of the thumbnail image relative to the original image. Maintians aspect ratio. default is 20.}

\item{overwrite}{Logical, passed to fs::file_copy. If \code{TRUE} then existing files with similar names will be written over. Default is \code{FALSE}.}
}
\value{
A folder named processed_photos in the data/processed/fulcrum directory. The folder contains full size sample photos renamed with strain names.
A thumbnails subfolder is also returned within the processed_photos folder. This folder contains image thumbnails. A dataframe identical to input \code{data}
with md5 hash values and file names for all photos.
}
\description{
\code{procPhotos} copies raw sample photos from the Fulcrum export and renames them to strain_name_C-label for use with CeNDR. The function also makes thumbnails.
}
