% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFulcrum.R
\name{readFulcrum}
\alias{readFulcrum}
\title{readFulcrum}
\usage{
readFulcrum(dir)
}
\arguments{
\item{dir}{The path to the base fulcrum project directory, dir/data/raw/fulcrum should contain:
nematode_field_sampling.csv,
nematode_field_sampling_sample_photo.csv,
nematode_isolation.csv,
nematode_isolation_s_labeled_plates.csv,
nematode_isolation_photos.csv

All five files are not required but it is recommended to use all exported files when possible.}
}
\value{
A list of dataframes named from the .csv files. For example, see the table below.
\tabular{ll}{
nematode_field_sampling \tab nematode_field_sampling.csv\cr
nematode_field_sampling_sample_photo \tab nematode_field_sampling_sample_photo.csv\cr
nematode_isolation \tab nematode_isolation.csv\cr
nematode_isolation_s_labeled_plates \tab nematode_isolation_s_labeled_plates.csv\cr
nematode_isolation_photos \tab nematode_isolation_photos.csv\cr
}
}
\description{
\code{readFulcrum} reads .csv files exported from Fulcrum into R
}
